1539. Kth Missing Positive Number
With this commit, I realized that I need to cut down on the time I spend on solving these problems on my own. I've decided to limit myself to only an hour (maximum) for any of these questions. After the time limit is reached, I get some help in order to learn, practice, and code the best-practice solutions and methods.